Refactored smells out of document/validation.rb
Smells reported by reek:

Before:

29: NoBrainer::Document::Validation::UniquenessValidator has no
    descriptive comment (IrresponsibleModule)
40: NoBrainer::Document::Validation::UniquenessValidator#validate_each
    doesn't depend on instance state (UtilityFunction)
40: NoBrainer::Document::Validation::UniquenessValidator#validate_each
    refers to document more than self (FeatureEnvy)

After:

Nothing to report

Notes:

The changes made for the reported smells were REALLY dumb. The
UniquenessValidator class *does* have a descriptive comment, but
because it was inside the class definition, reek didn't pick it up.

To remove the validate_each smells, the method was moved to the class
level, and the instance method calls the class method.