Refactored smells out of document/persistance.rb
Smells reported by reek:

Before:

29: NoBrainer::Document::Persistance#_create contains iterators
    nested 2 deep (NestedIterators)
49: NoBrainer::Document::Persistance#save contains iterators nested
    2 deep (NestedIterators)
47: NoBrainer::Document::Persistance#save has unused parameter
    'options' (UnusedParameters)

After:

38: NoBrainer::Document::Persistance#save has unused parameter
    'options' (UnusedParameters)

Notes:

Since part of this involved creating a _save method, and it was made
private, I've also moved the _create method to the private space. Also
extracted a private run_query method to get rid of the _create nested
iterators complaint.

Are there any plans to use the options parameter in #save, or could
that be removed?